Handle empty/corrupt FITS files

+++ rolodex/database.py
@@ -10,7 +10,7 @@ import multiprocessing as mp
 
 from census import find_all_datasets
 from db_tables import Base, Session, engine
-from db_tables import Files, Headers
+from db_tables import Files, Headers, Data
 from db_tables import Lampflash, Variability, Stims, Phd, Gain
 
 #-------------------------------------------------------------------------------
@@ -49,13 +49,13 @@ def insert_files(**kwargs):
             continue
 
         print("NEW: Found {}".format(full_filepath))
-        rootname = name.split('_')[0]
+        rootname = filename.split('_')[0]
         if not len(rootname) == 9:
             rootname = None
 
         session.add(Files(path=path,
                           name=filename,
-                          rootname=rootname)
+                          rootname=rootname))
 
         #-- Commit every 20 files to not lose too much progress if
         #-- a failure happens.
@@ -151,41 +151,47 @@ def update_header((args)):
     print(filename)
 
     session = Session()
-    with fits.open(filename) as hdu:
-        session.add(Headers(filetype=hdu[0].header['filetype'],
-                           instrume=hdu[0].header['instrume'],
-                           rootname=hdu[0].header['rootname'],
-                           imagetyp=hdu[0].header['imagetyp'],
-                           targname=hdu[0].header['targname'],
-                           ra_targ=hdu[0].header['ra_targ'],
-                           dec_targ=hdu[0].header['dec_targ'],
-                           proposid=hdu[0].header['proposid'],
-                           qualcom1=hdu[0].header.get('qualcom1', ''),
-                           qualcom2=hdu[0].header.get('qualcom2', ''),
-                           qualcom3=hdu[0].header.get('qualcom3', ''),
-                           quality=hdu[0].header.get('quality', ''),
-                           cal_ver=hdu[0].header['cal_ver'],
-                           opus_ver=hdu[0].header['opus_ver'],
-                           obstype=hdu[0].header['obstype'],
-                           obsmode=hdu[0].header['obsmode'],
-                           exptype=hdu[0].header['exptype'],
-                           detector=hdu[0].header['detector'],
-                           segment=hdu[0].header['segment'],
-                           detecthv=hdu[0].header['detecthv'],
-                           life_adj=hdu[0].header['life_adj'],
-                           fppos=hdu[0].header['fppos'],
-                           exp_num=hdu[0].header['exp_num'],
-                           cenwave=hdu[0].header['cenwave'],
-                           aperture=hdu[0].header['aperture'],
-                           opt_elem=hdu[0].header['opt_elem'],
-                           shutter=hdu[0].header['shutter'],
-                           extended=hdu[0].header['extended'],
-                           obset_id=hdu[0].header['obset_id'],
-                           asn_id=hdu[0].header['asn_id'],
-                           asn_tab=hdu[0].header['asn_tab'],
-                           file_id=f_key))
-        session.commit()
-        session.close()
+    try:
+        with fits.open(filename) as hdu:
+            session.add(Headers(filetype=hdu[0].header['filetype'],
+                                instrume=hdu[0].header['instrume'],
+                                rootname=hdu[0].header['rootname'],
+                                imagetyp=hdu[0].header['imagetyp'],
+                                targname=hdu[0].header['targname'],
+                                ra_targ=hdu[0].header['ra_targ'],
+                                dec_targ=hdu[0].header['dec_targ'],
+                                proposid=hdu[0].header['proposid'],
+                                qualcom1=hdu[0].header.get('qualcom1', ''),
+                                qualcom2=hdu[0].header.get('qualcom2', ''),
+                                qualcom3=hdu[0].header.get('qualcom3', ''),
+                                quality=hdu[0].header.get('quality', ''),
+                                cal_ver=hdu[0].header['cal_ver'],
+                                opus_ver=hdu[0].header['opus_ver'],
+                                obstype=hdu[0].header['obstype'],
+                                obsmode=hdu[0].header['obsmode'],
+                                exptype=hdu[0].header['exptype'],
+                                detector=hdu[0].header['detector'],
+                                segment=hdu[0].header['segment'],
+                                detecthv=hdu[0].header['detecthv'],
+                                life_adj=hdu[0].header['life_adj'],
+                                fppos=hdu[0].header['fppos'],
+                                exp_num=hdu[0].header['exp_num'],
+                                cenwave=hdu[0].header['cenwave'],
+                                aperture=hdu[0].header['aperture'],
+                                opt_elem=hdu[0].header['opt_elem'],
+                                shutter=hdu[0].header['shutter'],
+                                extended=hdu[0].header['extended'],
+                                obset_id=hdu[0].header['obset_id'],
+                                asn_id=hdu[0].header['asn_id'],
+                                asn_tab=hdu[0].header['asn_tab'],
+                                file_id=f_key))
+    except IOError as e:
+        print(e.message)
+        #-- Handle empty or corrupt FITS files
+        session.add(Headers(file_id=f_key))
+
+    session.commit()
+    session.close()
 
 #-------------------------------------------------------------------------------
 
@@ -215,50 +221,55 @@ def update_data((args)):
 
     session = Session()
 
-    with fits.open(filename) as hdu:
-        if len(hdu[1].data):
-            flux_mean=round(hdu[1].data['flux'].ravel().mean(), 5)
-            flux_max=round(hdu[1].data['flux'].ravel().max(), 5)
-            flux_std=round(hdu[1].data['flux'].ravel().std(), 5)
-        else:
-            flux_mean = None
-            flux_max = None
-            flux_std = None
+    try:
+        with fits.open(filename) as hdu:
+            if len(hdu[1].data):
+                flux_mean=round(hdu[1].data['flux'].ravel().mean(), 5)
+                flux_max=round(hdu[1].data['flux'].ravel().max(), 5)
+                flux_std=round(hdu[1].data['flux'].ravel().std(), 5)
+            else:
+                flux_mean = None
+                flux_max = None
+                flux_std = None
 
-        session.add(Data(hvlevela=hdu[1].header.get('hvlevela', -999),
-                         hvlevelb=hdu[1].header.get('hvlevelb', -999),
-                         date_obs=hdu[1].header['date-obs'],
-                         time_obs=hdu[1].header['time-obs'],
-                         expstart=hdu[1].header['expstart'],
-                         expend=hdu[1].header['expend'],
-                         exptime=hdu[1].header['exptime'],
-                         shift1a=hdu[1].header.get('shift1a', -999),
-                         shift1b=hdu[1].header.get('shift1b', -999),
-                         shift1c=hdu[1].header.get('shift1c', -999),
-                         shift2a=hdu[1].header.get('shift2a', -999),
-                         shift2b=hdu[1].header.get('shift2b', -999),
-                         shift2c=hdu[1].header.get('shift2c', -999),
-                         sp_loc_a=hdu[1].header.get('sp_loc_a', -999),
-                         sp_loc_b=hdu[1].header.get('sp_loc_b', -999),
-                         sp_loc_c=hdu[1].header.get('sp_loc_c', -999),
-                         sp_nom_a=hdu[1].header.get('sp_nom_a', -999),
-                         sp_nom_b=hdu[1].header.get('sp_nom_b', -999),
-                         sp_nom_c=hdu[1].header.get('sp_nom_c', -999),
-                         sp_off_a=hdu[1].header.get('sp_off_a', -999),
-                         sp_off_b=hdu[1].header.get('sp_off_b', -999),
-                         sp_off_c=hdu[1].header.get('sp_off_c', -999),
-                         sp_err_a=hdu[1].header.get('sp_err_a', -999),
-                         sp_err_b=hdu[1].header.get('sp_err_b', -999),
-                         sp_err_c=hdu[1].header.get('sp_err_c', -999),
+            session.add(Data(hvlevela=hdu[1].header.get('hvlevela', -999),
+                             hvlevelb=hdu[1].header.get('hvlevelb', -999),
+                             date_obs=hdu[1].header['date-obs'],
+                             time_obs=hdu[1].header['time-obs'],
+                             expstart=hdu[1].header['expstart'],
+                             expend=hdu[1].header['expend'],
+                             exptime=hdu[1].header['exptime'],
+                             shift1a=hdu[1].header.get('shift1a', -999),
+                             shift1b=hdu[1].header.get('shift1b', -999),
+                             shift1c=hdu[1].header.get('shift1c', -999),
+                             shift2a=hdu[1].header.get('shift2a', -999),
+                             shift2b=hdu[1].header.get('shift2b', -999),
+                             shift2c=hdu[1].header.get('shift2c', -999),
+                             sp_loc_a=hdu[1].header.get('sp_loc_a', -999),
+                             sp_loc_b=hdu[1].header.get('sp_loc_b', -999),
+                             sp_loc_c=hdu[1].header.get('sp_loc_c', -999),
+                             sp_nom_a=hdu[1].header.get('sp_nom_a', -999),
+                             sp_nom_b=hdu[1].header.get('sp_nom_b', -999),
+                             sp_nom_c=hdu[1].header.get('sp_nom_c', -999),
+                             sp_off_a=hdu[1].header.get('sp_off_a', -999),
+                             sp_off_b=hdu[1].header.get('sp_off_b', -999),
+                             sp_off_c=hdu[1].header.get('sp_off_c', -999),
+                             sp_err_a=hdu[1].header.get('sp_err_a', -999),
+                             sp_err_b=hdu[1].header.get('sp_err_b', -999),
+                             sp_err_c=hdu[1].header.get('sp_err_c', -999),
 
-                         flux_mean=flux_mean,
-                         flux_max=flux_max,
-                         flux_std=flux_std,
-                         file_id=f_key))
+                             flux_mean=flux_mean,
+                             flux_max=flux_max,
+                             flux_std=flux_std,
+                             file_id=f_key))
 
+    except IOError as e:
+        print(e.message)
+        #-- Handle empty or corrupt FITS files
+        session.add(Headers(file_id=f_key))
 
-        session.commit()
-        session.close()
+    session.commit()
+    session.close()
 
 #-------------------------------------------------------------------------------
 
@@ -297,6 +308,8 @@ def clear_all_databases():
 
     session = Session()
     for table in reversed(Base.metadata.sorted_tables):
+        if table.name == 'files':
+            continue
         try:
             print("Deleting {}".format(table.name))
             session.execute(table.delete())
@@ -311,14 +324,14 @@ if __name__ == "__main__":
     with open('../configure.yaml', 'r') as f:
         settings = yaml.load(f)
 
-    #clear_all_databases()
+    clear_all_databases()
 
     Base.metadata.create_all(engine)
     #Base.metadata.drop_all(engine, checkfirst=False)
     #Base.metadata.create_all(engine)
 
     print(settings)
-    insert_files(**settings)
-    #populate_primary_headers(settings['num_cpu'])
-    #populate_data(settings['num_cpu'])
+    #insert_files(**settings)
+    populate_primary_headers(settings['num_cpu'])
+    populate_data(settings['num_cpu'])
     #populate_lampflash()
